Solution to Count Leading Zeros in a Binary Number

+++ Count Leading Zeros in a Binary Number@@67882a731b4392d8fc1577f2/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main() {
-    int n;
-    scanf("%d",&n);
-    int k = 1,count=0;
+    long int n;
+    scanf("%ld",&n);
+    long int k = 1,count=0;
     while(k<n){
         k = k<<1;
         count++;
     }
-    printf("%d",32-count-1);
+    printf("%ld",32-count-1);
     return 0;
 }
